backend/transaction-fetcher/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.4
redis==5.0.1
requests==2.31.0
urllib3==2.1.0
websockets==12.0
kafka-python==2.0.2
